+++ CHANGELOG.md
@@ -7,7 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 https://github.com/nwn-dotnet/Anvil/compare/v8193.34.4...HEAD
 
 ### Added
-- N/A
+- Added `Color` equality members `Equals` and operators `==` `!=`
 
 ### Package Updates
 - Paket.Core 7.0.2 -> 7.1.4